examples(prompt_agents): consolidate agents/ + pipes/ into prompt_agents/
Merge `examples/agents/` and `examples/pipes/` into a single
`examples/prompt_agents/` folder. Both folders demoed the same product
concept (Prompt Agent target = hosted model + system prompt + optional
tool schema, with the runtime owning the tool-call loop). The split
between "building blocks" and "configs that use them" was artificial
for customers.

Changes:
- `git mv` all health_assistant tool modules, toolset YAMLs, pipeline
  configs, and the openclaw/ connector subfolder into prompt_agents/.
- Replace the two short folder READMEs with a single consolidated
  `examples/prompt_agents/README.md` matching the azure_doc_qa
  4-section structure: Scenario / Value-add / How to use / Behavior
  violation rate results.
- Update `tools.module` and `tools.toolset` paths in the moved YAMLs.
- Update path references in `CONFIG_REFERENCE.md`, root `README.md`,
  `examples/README.md`, `docs/targets/model-and-tools.md`,
  `scripts/README.md`, `p2m/cli.py`, and 5 test files.

Verification:
- `python -m compileall -q examples/prompt_agents` -> ok
- `pytest tests/test_tool_module_sandbox.py tests/test_openclaw_driver.py`
  `tests/test_inference_stage.py` -> 85 passed, 1 skipped
- `git grep -E 'examples[/\\]agents([/\\]|$)|examples[/\\]pipes([/\\]|$)|examples\.agents(\.|$)|examples\.pipes(\.|$)'` -> no hits in tracked files
