+++ examples/prompt_agents/health_assistant_generated_tools.yaml
