+++ examples/prompt_agents/health_assistant_tools.yaml
