+++ examples/prompt_agents/openclaw/SOUL.md
