+++ examples/prompt_agents/openclaw/docker-compose.yml
