+++ examples/prompt_agents/openclaw/entrypoint.sh
